Move template loading logic to a filter callback.
This moves all of the irving template logic to a callback on the 'wp_irving_components_route' hook. This will use the default WordPress template hierarchy to look for php or json files in a 'templates' folder inside your theme. The templates folder is filterable using the 'wp_irving_template_path' hook.

When the context is 'site' the tempalte loader will also attempt to load defaults from a defauls.php or defaults.json file.