Refactor context and registry into component instantiation
This reorganizes the Component class so that on instantiation, a Component class automatically
accepts configuration as a param and returns a hydrated instance of itself and its children.

To achieve this goal, depenencies to the context store and component registry have been moved into
the Component class, creating an intentional hard coupling between these parts of the architecture
and all classes have been moved into the Component namespace to better signal that coupling.

To let registered components to dynamically alter the configuration or children values during the
hydration process, the callback process has been split into two distinct callbacks:

* config_callback – A callback that accepts in the current config and returns a hydrated config array.
* children_callback – A callback that accepts the current array of children and config values and returns a hydrated array of children.

Continue refactor