Pass alias to Irving as a suggestion only.
This reverts previous behavior in the the `to_array()` method that would change the
name of a component to an alias, if set, and instead simply passes it to Irving as
a property which Irving should use as a hint for automatically mapping to another
component. Unit tests have all been updated to account for the change in json shape.