Update test to not rely on Component class
This modifies the expected value of the unit tests to not rely on the Component
class that we are testing in order to isolate only the output that we are testing
and avoid any side effects affecting our results.

This also registers the test provider explicitly in the method using the component so
the test passes in isolation, whereas it was previously failing in isolation but passing
during full test suite runs because test components from the component tests aren't being
properly cleaned up after those tests are run.