Add multi key context support
This allows a component to map properties of an object to specific config
values using the following notation:

```json
{
  "name": "example/component",
  "use_context": {
    "example/object": {
      "prop_foo": "config_key_bar",
    }
  }
}
```

It the above example, if `$object` is set as the value of the example/object
context. The example component will map the the value of `$object->prop_foo`
to it's own `config_key_bar` config value during hydration.